Fix behavior & logging on 'snap not found' error.
Previous behavior was just to raise and abort.
This behavior was masked in practice, because this error
(of a section file listing a snap name which doesn't exist
in the store) is masked if the snap name exists in the
cache, and ALL our snaps exist in the cache.

Without a cache file, this code will not run. Several
of the snap names in the .section files do not exist
in prod. (I don't yet understand how this could happen.
I guess the snaps were renamed?)

The new code logs a warning and then carries on
with the next snap.

Note also that the previous print call had a typo. Only one
'{}', but two args to .format().